Advanced_ai_model/Health_and_Fintness_ai/requirements.txt
streamlit==1.39.0
google-generativeai==0.8.3
Pillow==10.4.0
plotly==5.24.1
pandas==2.2.3